feat: remove userCache
